# avx2: Improve speed when > 10 input or output shards.