--- a/galois_gen_none.go
+++ b/galois_gen_none.go
@@ -5,6 +5,8 @@ package reedsolomon
 
 const maxAvx2Inputs = 0
 const maxAvx2Outputs = 0
+const minAvx2Size = 1
+const avxSizeMask = 0
 const avx2CodeGen = false
 
 func galMulSlicesAvx2(matrix []byte, in, out [][]byte, start, stop int) int {