… galoisAvx512_amd64_test.go
… galois_gen_amd64.go
… galois_gen_amd64.s
… reedsolomon_test.go